requirements.txt
pandas>=1.5.0
python-binance>=1.0.16
python-dateutil>=2.8.0
setuptools>=60.0.0
wheel>=0.37.0